WCAG Page Header fixes
* Added currently active icon menu to the service to ensure only one is open at a time.
* Changed the icon dropdown item layout to flexbox to resolve float issue in Firefox.
* Removed https from the webpack config and added `npm run start:https` task.
