+++ src/components/page-header/index.ts
@@ -3,3 +3,4 @@ export * from './page-header.component';
 export * from './navigation/navigation.component';
 export * from './icon-menu/icon-menu.component';
 export * from './custom-menu/custom-menu.directive';
+export * from './interfaces';
