+++ src/components/page-header/interfaces.ts
@@ -1,0 +1,16 @@
+export interface PageHeaderIconMenu {
+    icon: string;
+    label?: string;
+    badge?: number | string;
+    select?: (menu: PageHeaderIconMenu) => void;
+    dropdown?: PageHeaderIconMenuDropdownItem[];
+}
+
+export interface PageHeaderIconMenuDropdownItem {
+    icon?: string;
+    title: string;
+    subtitle?: string;
+    header?: boolean;
+    divider?: boolean;
+    select?: () => void;
+}
